# fact-electronica-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "FactElectronica",
    "version" : "1.0.0",
    "title" : "Client Fact Electronica",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost",
  "basePath" : "/DetalleFactura",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
    "/solicitud/solicitudInformacion" : {
      "post" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "addSolicitudInformacion",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "SolicitudInformacion by user",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SolicitudInformacion"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
    "/solicitud/solicitudInformacion/" : {
      "get" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "findAllSolicitudInformacion",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string",
            "default" : "available",
            "enum" : [ "available", "pending", "sold" ]
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SolicitudInformacion"
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
    "/solicitud/solicitudInformacion/{solicitudInformacionId}" : {
      "get" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "getSolicitudInformacionById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudInformacionId",
          "in" : "path",
          "description" : "ID of solicitudInformacion to return",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SolicitudInformacion"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "SolicitudInformacion not found"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "updateSolicitudInformacionWithForm",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudInformacionId",
          "in" : "path",
          "description" : "ID of solicitudInformacion that needs to be updated",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "name",
          "in" : "formData",
          "description" : "Updated name of the solicitudInformacion",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "status",
          "in" : "formData",
          "description" : "Updated status of the solicitudInformacion",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "updateSolicitudInformacion",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudInformacionId",
          "in" : "path",
          "description" : "SolicitudInformacion id to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Updated solicitudInformacion object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SolicitudInformacion"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid solicitudInformacion supplied"
          },
          "404" : {
            "description" : "SolicitudInformacion not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "deleteSolicitudInformacion",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Bearer",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "solicitudInformacionId",
          "in" : "path",
          "description" : "SolicitudInformacion id to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "SolicitudInformacion not found"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
    "/solicitud/solicitudInformacion/solicitudServicio/{solicitudServicioId}" : {
      "get" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "getSolicitudInformacionBysolicitudServicio",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudServicioId",
          "in" : "path",
          "description" : "ID of solicitudServicio to return",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SolicitudInformacion"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "SolicitudServicio not found"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "updateSolicitudInformacionWithSolicitudServicio",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudServicioId",
          "in" : "path",
          "description" : "ID of solicitudServicio that needs to be updated",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "name",
          "in" : "formData",
          "description" : "Updated name of the solicitudServicio",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "status",
          "in" : "formData",
          "description" : "Updated status of the solicitudServicio",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "updateSolicitudInformacionBySolicitudServicio",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solicitudServicioId",
          "in" : "path",
          "description" : "solicitudServicio id to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Updated solicitudInformacion object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SolicitudInformacion"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid solicitudServicio supplied"
          },
          "404" : {
            "description" : "SolicitudServicio not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "solicitudInformacion" ],
        "description" : "",
        "operationId" : "deleteSolicitudInformacionBySolicitudServicio",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Bearer",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "solicitudServicioId",
          "in" : "path",
          "description" : "SolicitudInformacion id to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "SolicitudServicio not found"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
    "/solicitud/solicitudServicio" : {
      "post" : {
        "tags" : [ "solicitudServicio" ],
        "description" : "",
        "operationId" : "addSolicitudServicio",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "SolicitudServicio by user",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SolicitudServicio"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
    "/solicitud/solicitudServicio/" : {
      "get" : {
        "tags" : [ "solicitudServicio" ],
        "description" : "",
        "operationId" : "findAllSolicitudServicio",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string",
            "default" : "available",
            "enum" : [ "available", "pending", "sold" ]
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SolicitudServicio"
              }
            }
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "ApiKeyAuth" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "ApiKeyAuth" : {
      "type" : "apiKey",
      "name" : "Bearer",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "SolicitudInformacion" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "solicitudServicio" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SolicitudServicio"
          }
        },
        "tipoSolicitud" : {
          "type" : "string"
        },
        "detalleSolInformacion" : {
          "type" : "string"
        },
        "notaDis" : {
          "type" : "string"
        },
        "notDes" : {
          "type" : "string"
        },
        "fechaBaja" : {
          "type" : "string"
        },
        "usuarioAct" : {
          "type" : "string"
        },
        "fechaAct" : {
          "type" : "string"
        },
        "solicitudesGeneradas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SolicitudGenerada"
          }
        },
        "cuentasSolicitud" : {
          "$ref" : "#/definitions/CuentasSolicitud"
        },
        "usuarioSolicitudInfo" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/UsuarioSolicitudInfo"
          }
        }
      },
      "xml" : {
        "name" : "SolicitudInformacion"
      }
    },
...
    "CuentasSolicitud" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "idSolInf" : {
          "type" : "integer",
          "format" : "int64"
        }
      },
      "xml" : {
        "name" : "CuentasSolicitud"
      }
    },
...
  },
...
}